+++ ChordKTV/Dtos/LrcLib/LrcLyricsDto.cs
@@ -1,7 +1,4 @@
-using System;
-
-namespace ChordKTV.Dtos;
-
+namespace ChordKTV.Dtos.LrcLib;
 public class LrcLyricsDto
 {
     public int Id { get; set; } // ID of original language song in LRCLIB
